Design Touch up for Signup. Prevent awkward overlays of values.
