+++ frontend/RaffleBay/RaffleBay/View/Profile View Elements /ProfileBidItemView.swift
